requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-ckeditor==6.7.2
django-js-asset==2.2.0
django-modeltranslation==0.19.11
idna==3.10
pillow==11.0.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
